feat: move Review and User type definitions to their respective files for better organization
