+++ src/type/common.ts
@@ -1,0 +1,6 @@
+export type searchParmas = {
+  className: string
+  page: number
+  sort: 'asc' | 'desc'
+  faculty: string
+}
