+++ src/type/review/index.ts
@@ -1,4 +1,6 @@
 import { ReviewDataSchema } from '@/type/review/schema'
+import { User } from '@/type/user'
+import { Likes } from '@prisma/client'
 import { z } from 'zod'
 
 export type ReviewFormState = {
@@ -11,6 +13,43 @@ export type ReviewFormState = {
     who?: string[]
   } | null
   message?: string | null
+}
+
+export type Review = {
+  faculty: string
+  id: number
+  date: string
+  title: string
+  star: number
+  createdBy: string
+  className: string
+  isAnonymous: boolean
+  evaluation: string
+  universityId: number
+  user: User
+  likes?: Likes[]
+}
+
+export type OriginalReview = {
+  id: number
+  date: string
+  className: string
+  title: string
+  star: number
+  evaluation: string
+  universityId: number
+  createdBy: string
+  faculty: string
+  isAnonymous: boolean
+}
+
+export type ReviewWithMetadata = Review & {
+  isLiked: boolean
+  isAuthor: boolean
+}
+
+export type ReviewWithIsLiked = Review & {
+  isLiked: boolean
 }
 
 export type ReviewData = z.infer<typeof ReviewDataSchema>
