+++ src/type/user/index.ts
@@ -1,0 +1,4 @@
+export type User = {
+  name: string | null
+  image: string | null
+}
